discord_bot/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
discord.py==2.4.0
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
ollama==0.3.1
sniffio==1.3.1
typing_extensions==4.12.2
yarl==1.9.4
